GCS - don't catch exceptions in 'exists?'
I think it makes more sense for the client to know about them.  An exception will either be due to a misconfiguration or credentials problem, which means no request will work and the app needs to be fixed, or else an intermittent network issue, which the client should be permitted to retry (if we see a lot of retries logged from the app, we might eventually put the exception-specific rescue back in here with some retry logic).